refactor status effect code
